Handle the case with low credit score
If the credit score of the applicant is low, and we cannot approve the
loan application for the requested amount and period. Then, we should
find the next lowest amount within the requested period and given credit
score. If that lowest number is below the global threshold (2000), then
find the next longest period we can approve for a global minimum
threshold.